review: Move row selection logic out of terminal rendering
Review records and displayed-row IDs now live in data.file_review.
Functions that decide which rows an action may select still live in
terminal output and are used by saved review state.

Those functions interpret change groups, selection reasons, and page
membership without producing terminal text. They belong with the review
data that they examine.

This commit moves selectable-row calculations into data.file_review and
redirects rendering and saved-state code to that module.

The next commit will guard the new imports before default review-page
selection also moves out of terminal output.
